Add scala doc to generated grpc proto file
This is a draft of the full implementation.

Parse the scala doc from generated Scala doc HTML file. So the project
using this should run `sbt doc` first before run the generation tasks.
Will auto include it in the task dependency in the future.

Only implemented scala doc on service and method for now. Scala doc
dicretory is also hard coded.